python/kwiver/requirements.txt
numpy==1.13.0
pillow==7.1.0
six==1.10.0

# For Testing
nose==1.3.7
coverage==4.4.1
